Update wiremock asserts
